% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{browse}
\alias{browse}
\title{Browse the joinpoint files}
\usage{
browse(jp)
}
\arguments{
\item{jp}{the joinpoint analysis to browse}
}
\description{
Open the directory in temporary files where joinpoint files are stored
}
\examples{
jp = joinpoint_example()
print(dir)
#browseURL(dir)
}
