% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_export.R
\name{export_options}
\alias{export_options}
\title{Export options}
\source{
\link{https://surveillance.cancer.gov/joinpoint/Joinpoint_Help_4.8.0.1.pdf}
}
\usage{
export_options(
  models = c("best fit", "all"),
  remove_jp_flags = TRUE,
  remove_best_fit_flags = FALSE,
  output_bygroup_headers = FALSE,
  all_models_in_same_column = FALSE,
  include_jp_estimates = FALSE,
  include_apcs = TRUE,
  x_precision = 9,
  y_precision = 9,
  model_precision = 3,
  estimated_joinpoint_precision = 3,
  regression_coefficients_precision = 3,
  covariance_matrix_precision = 3,
  correlation_matrix_precision = 3,
  apc_precision = 3,
  aapc_precision = 3,
  aapc_segemnt_ranges_precision = 3,
  pvalue_precision = 3,
  aapc_full_range = FALSE,
  aapc_start_range1 = NULL,
  aapc_end_range1 = NULL,
  aapc_start_range2 = NULL,
  aapc_end_range2 = NULL,
  aapc_start_range3 = NULL,
  aapc_end_range3 = NULL,
  aapc_last_obs = FALSE,
  export_bad_cohorts = TRUE,
  export_report = TRUE,
  export_data = TRUE,
  export_apc = TRUE,
  export_aapc = TRUE,
  export_ftest = TRUE,
  export_pairwise = TRUE,
  export_jump_cr = TRUE
)
}
\arguments{
\item{models}{one of \code{c("best fit", "all")}. Using \code{all} might be insightful but rather messes with the output.}

\item{remove_jp_flags}{remove_jp_flags}

\item{remove_best_fit_flags}{removes the \code{final_selected_model} column}

\item{output_bygroup_headers}{should only be used if \code{models==best}. If \code{FALSE}, result table has \code{model}, \code{apc}, \code{joinpoints}. If \code{TRUE}, result table has \code{jp2}, \code{apc2} (if selected model is with 2 joinpoints) but rather messes with the output.}

\item{all_models_in_same_column}{all_models_in_same_column}

\item{include_jp_estimates}{include_jp_estimates}

\item{include_apcs}{include_apcs}

\item{x_precision}{x_precision}

\item{y_precision}{y_precision}

\item{model_precision}{model_precision}

\item{estimated_joinpoint_precision}{estimated_joinpoint_precision}

\item{regression_coefficients_precision}{regression_coefficients_precision}

\item{covariance_matrix_precision}{covariance_matrix_precision}

\item{correlation_matrix_precision}{correlation_matrix_precision}

\item{apc_precision}{apc_precision}

\item{aapc_precision}{aapc_precision}

\item{aapc_segemnt_ranges_precision}{aapc_segemnt_ranges_precision}

\item{pvalue_precision}{pvalue_precision}

\item{aapc_full_range}{aapc_full_range}

\item{aapc_start_range1}{aapc_start_range1}

\item{aapc_end_range1}{aapc_end_range1}

\item{aapc_start_range2}{aapc_start_range2}

\item{aapc_end_range2}{aapc_end_range2}

\item{aapc_start_range3}{aapc_start_range3}

\item{aapc_end_range3}{aapc_end_range3}

\item{aapc_last_obs}{aapc_last_obs}

\item{export_bad_cohorts}{export_bad_cohorts}

\item{export_report}{export_report}

\item{export_data}{export_data}

\item{export_apc}{export_apc}

\item{export_aapc}{export_aapc}

\item{export_ftest}{export_ftest}

\item{export_pairwise}{export_pairwise}

\item{export_jump_cr}{export_jump_cr}
}
\value{
a string (character of length 1) representing the .ini file
}
\description{
Create a character representation of the \code{.ini} file needed for export options.
}
