% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tempdir}
\alias{get_tempdir}
\title{Temporary directory for joinpoint files}
\usage{
get_tempdir()
}
\description{
Creates a timestamped directory in Local/Temp
}
\examples{
dir = get_tempdir()
print(dir)
#browseURL(dir)
}
