% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinpoint.R
\name{joinpoint}
\alias{joinpoint}
\title{API to run a joinpoint model}
\usage{
joinpoint(
  data,
  x,
  y,
  by = NULL,
  se = NULL,
  y_type = c("Age-Adjusted Rate", "Crude rate", "Percent", "Proportion", "Count"),
  export_opt_ini = export_options(),
  run_opt_ini = run_options(),
  cmd_path = getOption("joinpoint_path",
    "C:/Program Files (x86)/Joinpoint Command/jpCommand.exe"),
  dir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{data}{A data frame}

\item{x}{\verb{<tidy-select>} the independant variable (for instance the year)}

\item{y}{\verb{<tidy-select>} the dependant variable of type \code{y_type}}

\item{by}{\verb{<tidy-select>} one or several stratification variable (for instance sex)}

\item{se}{\verb{<tidy-select>} the standard error of the dependant variable. Can be left as \code{NULL} at the cost of a longer computation. See https://seer.cancer.gov/seerstat/WebHelp/Rate_Algorithms.htm for calculation formulas.}

\item{y_type}{the type of dependant variable. Must be one of \code{c("Age-Adjusted Rate", "Crude rate", "Percent", "Proportion", "Count")}.}

\item{export_opt_ini}{the result of \code{\link[=export_options]{export_options()}}}

\item{run_opt_ini}{the result of \code{\link[=run_options]{run_options()}}}

\item{cmd_path}{the path to the executable. Can usualy be left default to \code{"C:/Program Files (x86)/Joinpoint Command/jpCommand.exe"}. Can also be set through \code{options(joinpoint_path="my/path/to/jp.exe")}.}

\item{dir}{The temporary directory where all the temporary files will be written}

\item{verbose}{unused for the moment}
}
\value{
the list of the output tables
}
\description{
Use the Command-Line version of "Joinpoint Regression Software" provided by the NIH to run a regression.
The software must be downloaded at https://surveillance.cancer.gov/joinpoint/callable/ and installed on a \strong{Windows} computer. I am not aware of a version of this software for Linux or MacOS.
}
\details{
This function will generate the \code{.ini} files, run the software, and then parse the result files as an plain old R list.
}
