% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jp_plot.R
\name{jp_plot}
\alias{jp_plot}
\title{Title}
\usage{
jp_plot(
  jp,
  legend_pattern = getOption("jp_plot_pattern", "{xmin}-{xmax}: {slope}"),
  by_level = NULL
)
}
\arguments{
\item{jp}{A list generated using \code{\link[=joinpoint]{joinpoint()}}.}

\item{legend_pattern}{\code{\link[glue:glue]{glue::glue()}} pattern for the legend. Can use variables \code{slope}, \code{xmin}, and \code{xmax}. Can be set through options, e.g. \code{options(jp_plot_pattern="-{slope}-")}.}

\item{by_level}{One or several stratification levels. Works only if \code{jp} was made using one single stratification variable.}
}
\value{
\code{patchwork} of \code{ggplot}s
}
\description{
Title
}
