% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{nih_sample_data}
\alias{nih_sample_data}
\title{Sample dataset}
\format{
A dataframe with 105 observations on 4 variables
}
\source{
Dataset that comes with NIH's Joinpoint software
}
\usage{
nih_sample_data
}
\description{
Sample dataset
}
\keyword{datasets}
