% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_run.R
\name{run_options}
\alias{run_options}
\title{Design options}
\usage{
run_options(
  model = c("linear", "ln"),
  data_shift = 0,
  min_joinpoints = 0,
  max_joinpoints = 4,
  pairwise = NULL,
  pairwise_signif_lvl = 0.05,
  pairwise_n_permut = 4499,
  method = c("grid", "hudsons"),
  min_obs_end = 2,
  min_obs_between = 2,
  n_obs_between = 0,
  model_selection_method = c("permutation test", "bic", "mbic",
    "data dependent selection", "wbic", "wbic-alt"),
  permutation_signif_lvl = 0.05,
  n_permutations = 4499,
  early_stopping = c("b-value", "curtailed", "fixed"),
  run_type = c("calculated", "provided"),
  rates_per_n = 1e+05,
  dependent_variable_type = c("count", "crude rate", "age-adjusted rate", "proportion",
    "percent"),
  het_error = c("constant variance", "standard error", "poisson rate", "poisson count"),
  het_error_var_location = NULL,
  ci_method = c("parametric", "empirical quantile method 1",
    "empirical quantile method 2"),
  n_cores = 1,
  delay_type = c("delay", "non-delay", "both"),
  autocorr_errors = c("number", "estimated"),
  jump_model = FALSE,
  comparability_ratio = FALSE,
  include_std_analysis = FALSE,
  jump_location = 9999,
  comparability_ratio_value = 0,
  cr_variance = 0,
  joinpoint_alpha_lvl = 0.05,
  apc_alpha_lvl = 0.05,
  aapc_alpha_lvl = 0.05,
  jump_cr_alpha_lvl = 0.05,
  random_seed = 7160,
  empirical_quantile_seed = 10000,
  empirical_quantile_seed_type = c("constant", "varying"),
  n_resample = 1000,
  madwd = FALSE,
  madwd_psi = 0
)
}
\arguments{
\item{model}{one of \code{c("linear", "ln")}}

\item{data_shift}{data_shift}

\item{min_joinpoints}{min_joinpoints}

\item{max_joinpoints}{max_joinpoints}

\item{pairwise}{pairwise}

\item{pairwise_signif_lvl}{pairwise_signif_lvl}

\item{pairwise_n_permut}{pairwise_n_permut}

\item{method}{method}

\item{min_obs_end}{min_obs_end}

\item{min_obs_between}{min_obs_between}

\item{n_obs_between}{n_obs_between}

\item{model_selection_method}{model_selection_method}

\item{permutation_signif_lvl}{permutation_signif_lvl}

\item{n_permutations}{n_permutations}

\item{early_stopping}{early_stopping}

\item{run_type}{run_type}

\item{rates_per_n}{rates_per_n}

\item{dependent_variable_type}{dependent_variable_type}

\item{het_error}{het_error}

\item{het_error_var_location}{het_error_var_location}

\item{ci_method}{ci_method}

\item{n_cores}{n_cores}

\item{delay_type}{delay_type}

\item{autocorr_errors}{autocorr_errors}

\item{jump_model}{jump_model}

\item{comparability_ratio}{comparability_ratio}

\item{include_std_analysis}{include_std_analysis}

\item{jump_location}{jump_location}

\item{comparability_ratio_value}{comparability_ratio_value}

\item{cr_variance}{cr_variance}

\item{joinpoint_alpha_lvl}{joinpoint_alpha_lvl}

\item{apc_alpha_lvl}{apc_alpha_lvl}

\item{aapc_alpha_lvl}{aapc_alpha_lvl}

\item{jump_cr_alpha_lvl}{jump_cr_alpha_lvl}

\item{random_seed}{random_seed}

\item{empirical_quantile_seed}{empirical_quantile_seed}

\item{empirical_quantile_seed_type}{empirical_quantile_seed_type}

\item{n_resample}{n_resample}

\item{madwd}{madwd}

\item{madwd_psi}{madwd_psi}
}
\value{
a string (character of length 1) representing the .ini file
}
\description{
Create a character representation of the \code{.ini} file needed for run options.
}
